chore(phpstan): raise static analysis to level 2 (429→0, fix-forward) (#3559)
* chore(phpstan): systemic groundwork for level 2 (429→147 errors)

Centralized fixes that clear ~66% of level-2 findings with no per-call-site churn:

- DispatchesEvents trait: set_class_context()/get_function_context() private→protected
  static. The trait is mixed into ~60 classes, so static:: calls to a private method
  produced one staticClassAccess.privateMethod per using class (~112 reports, 1 defect).
- Carbon date/time macros: resolved via Carbon's official PHPStan MacroExtension
  (vendor extension.neon) fed by a mixin registration in .phpstan/bootstrap.php, instead
  of stubbing Carbon\CarbonInterface (which would clobber Carbon's own @method API).
- laravel-gaps.stub: declares concrete Laravel methods our code calls through looser
  contracts (ConnectionInterface::getDriverName, Filesystem::url, Auth Factory/Guard,
  Foundation\Application, Socialite Provider) — no Larastan, so vanilla PHPStan needs these.
- Language.php: static:: → self:: for private constants (class is final-by-usage, not extended).
- AuthenticateSession::guard() returns $this->auth->guard() (the actual Guard) instead of
  the AuthManager, matching upstream Laravel and keeping the return type honest.

Config still at level 1; this only adds stubs/extension so the level-2 bump lands clean.


* fix(phpstan): real bugs surfaced by level-2 analysis (147→132)

PHPStan level 2 flagged genuine latent bugs (not just type noise):

- Users/Import: session()->exist(...) → exists() (the typo'd method does not exist;
  would fatal at runtime on the LDAP-import path).
- DateTimeHelper: add parseDb24hTime() — Format's FromFormat::Db24hTime case called a
  method that never existed (runtime fatal on that path). Mirrors parseUser24hTime in dbTimezone.
- Cast: narrow ReflectionType→ReflectionNamedType before ->getName() (union/intersection types
  have no getName()).
- Projects::getProjectProgress: cast DateInterval::format('%a') (string) to int before division.
- FileManager::convertPHPSizeToBytes: cast substr() to int before *= (string arithmetic).
- ApiGuard: gate getAPIKey() behind instanceof ApiRequest (getAPIKey lives on ApiRequest,
  not the IncomingRequest base; isApiRequest() is URI-based and doesn't prove the instance type).
- Install: Cache::store('installation')->flush() → clear() (flush() isn't on the Repository
  contract; clear() is the PSR-16 method that wipes the store).
- Setting/FileManagerInterface: declare getFileUrl() on the interface (concrete already has it).
- ExceptionHandler::isHttpException: @phpstan-assert-if-true so getStatusCode()/getHeaders()
  narrow correctly.
- Connector\Providers::getProvider(): return type provider-model → object (it returns provider
  service objects with connect()/getEntities(), never the model — the hint was wrong).
- laravel-gaps.stub: Carbon::addRealMinutes (Carbon-3 magic method, annotation dropped upstream).


* fix(phpstan): resolve remaining level-2 logic bugs (132→112)

return.missing — add explicit `return null` where a typed method fell through:
EventDispatcher::dispatch, ApiGuard::id, WebGuard::id, Projects::getProjectName,
AuthenticateSession::redirectTo.

method.void — Users\Repositories::setPicture void→bool (+return true): it was used in
setProfilePicture's `&& $this->userRepo->setPicture(...) &&` chain, so the void/null return
made the branch ALWAYS falsy and the old profile picture was never deleted. Queue::addToQueue
/addJob no longer `return` the void addMessageToQueue() result.

method/property.nonObject — CheckPermissions narrows $request->route() with `instanceof Route`
before getControllerClass()/getActionMethod(); Integration guards `is_object()` before ->id.

arguments.count — extra args were silently dropped by PHP at runtime; removed to match the real
signatures (FLAGGED for review where a feature was being lost):
- Entityrelations saveRelationship/getRelationshipByEntity: passed 6/4 args to Setting's 2-arg
  saveSetting/getSetting — the methods are broken (relationship data dropped); trimmed + FIXME.
- Files upload: passed the md5 filename as the disk arg (would fail disk lookup); now upload($file)
  + FIXME on the encName mismatch.
- getProjectsUserHasAccessTo('open') ×3 and getAllClientsAvailableToUser(...,$client): dropped
  status/client filters that the method never accepted.
- quickAddTicket(...,projectId) and parseUserDateTime(...,FromFormat) extra args removed.

method.notFound — CarbonMacros setToUserTimezone/setToDbTimezone closures now return
CarbonImmutable (not CarbonInterface), so chained macro calls keep resolving.


* fix(phpstan): property/method access bugs at level 2 (112→~90)

- Widgets\Models\Widget: declare `public bool $isNew` — the Widgets service set $widget->isNew
  dynamically (deprecated dynamic property in PHP 8.2+).
- Auth ApiGuard::id/WebGuard::id: use the Authenticatable contract's getAuthIdentifier()
  instead of ->id (the contract has no $id property).
- FacadeRule (our own PHPStan rule): PhpParser\Node\Name->parts → ->getParts() (the $parts
  property was removed in nikic/php-parser v5).
- Projects::notifyProjectUsers: @var Notification after dispatch_filter() (which returns mixed),
  restoring the type so ->module resolves.
- Timesheets: @var CarbonImmutable on a chain ending in startOfDay() (Carbon annotates it as
  returning the interface, where the macro can't resolve).


* chore(phpstan): mechanical docblock/namespace sweep + bump to level 2

Clears the remaining 102 level-2 findings across 47 files — all docblock / type-annotation /
namespace fixes, no runtime logic changes:

- parameter.notFound: stale @param names dropped/renamed to match real signatures.
- return/parameter/property.phpDocType: @return/@param/@var types reconciled with native types
  (or removed where redundant).
- class.notFound: @var \Leantime\Core\Bootstrap\Application → \Leantime\Core\Application (the
  Bootstrap segment never existed) in HandleExceptions/ExceptionHandler.
- throws.notThrowable: corrected exception FQCNs — \Illuminate\Contracts\Container\
  BindingResolutionException and \Illuminate\Http\Exceptions\HttpResponseException.
- phpDoc.parseError / varTag.variableNotFound / class.nameCase: malformed tags repaired,
  stale inline @var removed, Environment case corrected.
- laravel-gaps.stub: Cache\Repository::lock() (forwarded to a LockProvider store via __call;
  used by StartSession session locking).

`.phpstan/phpstan.neon` level 1 → 2. `make phpstan` is green at level 2 (838 files, 0 errors).


* style(phpstan): apply Laravel Pint to the level-2 changes

Pint auto-fixes on the files touched by the level-2 work: drop superfluous/empty PHPDoc
tags, remove the now-unused Carbon\CarbonInterface import in CarbonMacros, and normalize
spacing/braces. `pint --test` is clean (628 files) and `make phpstan` is still green at level 2.


* fix(phpstan): address Copilot review on the level-2 PR

- ApiGuard: give $apiKey a '' default so empty($this->apiKey) is safe when the guard is
  constructed for a non-API request (typed property was never initialized otherwise).
- ApiGuard::id / WebGuard::id: both guards' user() return a stdClass at runtime (the (object)
  cast / AuthUser::retrieveById), so calling getAuthIdentifier() would fatal. Read ->id off the
  object with an accurate @var instead. (The real Authenticatable conversion is the level-3 pass.)
- Users\Repositories::setPicture: return the actual update result (> 0) instead of always true,
  so the caller's control flow reflects whether the row was updated.
- DefaultConfig: spelling, identifieble → identifiable.

phpstan level 2 still green; pint clean.


---------

+++ app/Core/Files/FileManager.php
@@ -430,7 +430,7 @@ class FileManager implements FileManagerInterface
         if (! in_array($sSuffix, ['P', 'T', 'G', 'M', 'K'])) {
             return (int) $sSize;
         }
-        $iValue = substr($sSize, 0, -1);
+        $iValue = (int) substr($sSize, 0, -1);
         switch ($sSuffix) {
             case 'P':
                 $iValue *= 1024;
